@@ -11,65 +11,65 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.0",
-    "@angular/common": "~12.2.0",
-    "@angular/compiler": "~12.2.0",
-    "@angular/core": "~12.2.0",
-    "@angular/forms": "~12.2.0",
-    "@angular/localize": "~12.2.0",
-    "@angular/platform-browser": "~12.2.0",
-    "@angular/platform-browser-dynamic": "~12.2.0",
-    "@angular/router": "~12.2.0",
-    "@angular/service-worker": "~12.2.0",
-    "@ctrl/ngx-codemirror": "^5.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^10.0.0",
-    "angular-gridster2": "^12.1.0",
-    "angular-split": "^5.0.0",
-    "bootstrap": "~4.6.0",
+    "@angular/animations": "~13.1.1",
+    "@angular/cdk": "~13.1.1",
+    "@angular/common": "~13.1.1",
+    "@angular/compiler": "~13.1.1",
+    "@angular/core": "~13.1.1",
+    "@angular/forms": "~13.1.1",
+    "@angular/localize": "~13.1.1",
+    "@angular/platform-browser": "~13.1.1",
+    "@angular/platform-browser-dynamic": "~13.1.1",
+    "@angular/router": "~13.1.1",
+    "@angular/service-worker": "~13.1.1",
+    "@ctrl/ngx-codemirror": "~5.1.1",
+    "@ng-bootstrap/ng-bootstrap": "^11.0.0",
+    "angular-gridster2": "^13.1.0",
+    "angular-split": "^13.1.0",
+    "bootstrap": "~4.6.1",
     "bootstrap-darkmode": "^0.9.1",
-    "bootstrap-icons": "^1.6.1",
+    "bootstrap-icons": "^1.7.2",
     "bson-objectid": "^2.0.1",
-    "codemirror": "^5.62.2",
+    "codemirror": "^5.65.0",
     "file-saver": "^2.0.5",
-    "highlight.js": "^11.2.0",
-    "keycloak-angular": "^8.3.0",
-    "keycloak-js": "12.0.3",
+    "highlight.js": "^11.3.1",
+    "keycloak-angular": "^9.0.0",
+    "keycloak-js": "15.1.0",
     "ng-bootstrap-darkmode": "^0.5.0",
-    "ng-terminal": "^3.0.1",
-    "ng2-dragula": "^2.1.1",
-    "ngx-clipboard": "^14.0.1",
-    "ngx-drag-drop": "^2.0.0",
-    "rxjs": "~6.6.7",
-    "tslib": "^2.3.0",
+    "ng-terminal": "^4.0.0-alpha.2",
+    "ngx-clipboard": "^15.0.1",
+    "ngx-drag-drop": "^13.1.0-alpha.1",
+    "rxjs": "~7.4.0",
+    "tslib": "^2.3.1",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.0",
-    "@angular-eslint/builder": "~12.3.1",
-    "@angular-eslint/eslint-plugin": "~12.3.1",
-    "@angular-eslint/eslint-plugin-template": "~12.3.1",
-    "@angular-eslint/schematics": "~12.3.1",
-    "@angular-eslint/template-parser": "~12.3.1",
-    "@angular/cli": "~12.2.0",
-    "@angular/compiler-cli": "~12.2.0",
-    "@angular/language-service": "~12.2.0",
-    "@types/codemirror": "^5.60.2",
-    "@types/file-saver": "^2.0.3",
+    "@angular-devkit/build-angular": "~13.1.2",
+    "@angular-eslint/builder": "~13.0.1",
+    "@angular-eslint/eslint-plugin": "~13.0.1",
+    "@angular-eslint/eslint-plugin-template": "~13.0.1",
+    "@angular-eslint/schematics": "~13.0.1",
+    "@angular-eslint/template-parser": "~13.0.1",
+    "@angular/cli": "~13.1.2",
+    "@angular/compiler-cli": "~13.1.1",
+    "@angular/language-service": "~13.1.1",
+    "@types/codemirror": "^5.60.5",
+    "@types/file-saver": "^2.0.4",
     "@types/jasmine": "~3.6.11",
     "@types/jasminewd2": "~2.0.10",
-    "@types/node": "^16.4.13",
-    "@typescript-eslint/eslint-plugin": "^4.29.1",
-    "@typescript-eslint/parser": "^4.29.1",
-    "eslint": "^7.32.0",
+    "@types/node": "^16.11.17",
+    "@typescript-eslint/eslint-plugin": "^5.8.1",
+    "@typescript-eslint/parser": "^5.8.1",
+    "eslint": "^8.5.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.2",
-    "karma": "~6.3.4",
+    "karma": "~6.3.9",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.3",
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "~7.0.0",
-    "ts-node": "^10.2.0",
-    "typescript": "~4.3.5"
+    "ts-node": "^10.4.0",
+    "typescript": "~4.5.4"
   }
 }
