@@ -20,55 +20,63 @@
     "test:e2e": "jest --config ./apps/services/test/jest-e2e.json"
   },
   "dependencies": {
-    "@nestjs/axios": "^0.0.2",
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/core": "^8.0.0",
+    "@elastic/elasticsearch": "~7.13.0",
+    "@nestjs/axios": "^0.0.3",
+    "@nestjs/common": "^8.2.3",
+    "@nestjs/core": "^8.2.3",
+    "@nestjs/elasticsearch": "^8.0.0",
     "@nestjs/event-emitter": "^1.0.0",
     "@nestjs/jwt": "^8.0.0",
     "@nestjs/microservices": "^8.1.2",
-    "@nestjs/mongoose": "^8.0.1",
+    "@nestjs/mongoose": "^9.0.1",
     "@nestjs/passport": "^8.0.1",
-    "@nestjs/platform-express": "^8.0.0",
-    "@nestjs/schedule": "^1.0.1",
-    "@nestjs/swagger": "^5.0.9",
-    "@types/dockerode": "^3.2.7",
+    "@nestjs/platform-express": "^8.2.3",
+    "@nestjs/schedule": "^1.0.2",
+    "@nestjs/swagger": "^5.1.5",
+    "@types/dockerode": "^3.3.0",
     "@types/passport-jwt": "^3.0.6",
-    "axios": "0.21.4",
+    "axios": "^0.24.0",
     "bson-objectid": "^2.0.1",
     "class-transformer": "^0.4.0",
-    "class-validator": "^0.13.1",
-    "dockerode": "^3.3.0",
-    "mongoose": "^6.0.3",
+    "class-validator": "^0.13.2",
+    "dockerode": "^3.3.1",
+    "glob-to-regexp": "^0.4.1",
+    "mongoose": "^6.1.2",
+    "multer": "^1.4.4",
     "nats": "^2.2.0",
-    "passport": "^0.4.1",
+    "passport": "^0.5.0",
     "passport-jwt": "^4.0.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
-    "rxjs": "^7.2.0",
-    "swagger-ui-express": "^4.1.6"
+    "rxjs": "^7.4.0",
+    "swagger-ui-express": "^4.2.0",
+    "unzipper": "^0.10.11"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.0.0",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
+    "@nestjs/cli": "^8.1.6",
+    "@nestjs/schematics": "^8.0.5",
+    "@nestjs/testing": "^8.2.3",
     "@types/cron": "^1.7.3",
     "@types/express": "^4.17.13",
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.0.0",
+    "@types/glob-to-regexp": "^0.4.1",
+    "@types/jest": "^27.0.3",
+    "@types/multer": "^1.4.7",
+    "@types/node": "^16.11.13",
     "@types/supertest": "^2.0.11",
-    "@typescript-eslint/eslint-plugin": "^4.28.2",
-    "@typescript-eslint/parser": "^4.28.2",
-    "eslint": "^7.30.0",
+    "@types/unzipper": "^0.10.5",
+    "@typescript-eslint/eslint-plugin": "^5.7.0",
+    "@typescript-eslint/parser": "^5.7.0",
+    "eslint": "^8.4.1",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "jest": "^27.0.6",
-    "prettier": "^2.3.2",
-    "supertest": "^6.1.3",
-    "ts-jest": "^27.0.3",
-    "ts-loader": "^9.2.3",
-    "ts-node": "^10.0.0",
-    "tsconfig-paths": "^3.10.1",
-    "typescript": "^4.3.5"
+    "eslint-plugin-prettier": "^4.0.0",
+    "jest": "^27.4.5",
+    "prettier": "^2.5.1",
+    "supertest": "^6.1.6",
+    "ts-jest": "^27.1.1",
+    "ts-loader": "^9.2.6",
+    "ts-node": "^10.4.0",
+    "tsconfig-paths": "^3.12.0",
+    "typescript": "^4.5.4"
   },
   "jest": {
     "moduleFileExtensions": [
