@@ -2,12 +2,12 @@ import {HttpClientModule} from '@angular/common/http';
 import {APP_INITIALIZER, NgModule} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {BrowserModule} from '@angular/platform-browser';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {ServiceWorkerModule} from '@angular/service-worker';
 
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 import {KeycloakAngularModule, KeycloakService} from 'keycloak-angular';
 import {NgBootstrapDarkmodeModule, THEME_LOADER, THEME_SAVER, ThemeLoader, ThemeSaver} from 'ng-bootstrap-darkmode';
-import {DragulaModule} from 'ng2-dragula';
 import {of} from 'rxjs';
 
 import {environment} from '../environments/environment';
@@ -23,6 +23,8 @@ import {PrivacyService} from './privacy.service';
 import {PrivacyComponent} from './privacy/privacy.component';
 import {SharedModule} from './shared/shared.module';
 import {UserModule} from './user/user.module';
+import { ToastListComponent } from './toast-list/toast-list.component';
+import { HomeComponent } from './home/home.component';
 
 function initializeKeycloak(keycloak: KeycloakService) {
   return () => keycloak.init({
@@ -52,14 +54,16 @@ function initializeKeycloak(keycloak: KeycloakService) {
     FeedbackComponent,
     PrivacyComponent,
     AboutComponent,
+    ToastListComponent,
+    HomeComponent,
   ],
   imports: [
     BrowserModule,
     FormsModule,
     HttpClientModule,
+    BrowserAnimationsModule,
     NgbModule,
     NgBootstrapDarkmodeModule,
-    DragulaModule.forRoot(),
     KeycloakAngularModule,
     SharedModule,
     AppRoutingModule,
