@@ -1,47 +1,38 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {EvaluationModalComponent} from './pages/evaluation-modal/evaluation-modal.component';
+
 import {CourseComponent} from './pages/course/course.component';
 import {CreateCourseComponent} from './pages/create-course/create-course.component';
 import {CreateSolutionComponent} from './pages/create-solution/create-solution.component';
-import {EditAssignmentComponent} from './pages/edit-assignment/edit-assignment.component';
-import {EditTaskModalComponent} from './pages/edit-task-modal/edit-task-modal.component';
 import {MyAssignmentsComponent} from './pages/my-assignments/my-assignments.component';
 import {MyCoursesComponent} from './pages/my-courses/my-courses.component';
 import {MySolutionsComponent} from './pages/my-solutions/my-solutions.component';
-import {SolutionTableComponent} from './pages/solution-table/solution-table.component';
-import {SolutionComponent} from './pages/solution/solution.component';
 
 const routes: Routes = [
-  {path: '', component: MyAssignmentsComponent},
-  {path: 'solutions', component: MySolutionsComponent},
-  {path: 'courses', component: MyCoursesComponent},
+  {path: '', component: MyAssignmentsComponent, data: {title: 'My Assignments'}},
+  {path: 'solutions', component: MySolutionsComponent, data: {title: 'My Solutions'}},
+  {path: 'courses', component: MyCoursesComponent, data: {title: 'My Courses'}},
+  {path: 'courses/create', component: CreateCourseComponent, data: {title: 'Create Course'}},
+  {path: 'courses/:cid', component: CourseComponent, data: {title: 'Course'}},
+  {path: 'courses/:cid/assignments/:aid', component: CourseComponent, data: {title: 'Course'}},
   {
     path: 'create',
-    component: EditAssignmentComponent,
-    children: [
-      {path: 'tasks/:task', component: EditTaskModalComponent},
-    ],
+    loadChildren: () => import('./modules/edit-assignment/edit-assignment.module').then(m => m.EditAssignmentModule),
   },
-  {path: ':aid', component: CreateSolutionComponent},
-  {
-    path: ':aid/edit',
-    component: EditAssignmentComponent,
-    children: [
-      {path: 'tasks/:task', component: EditTaskModalComponent},
-    ],
-  },
-  {path: ':aid/solutions', component: SolutionTableComponent},
+  {path: ':aid/solutions/new', component: CreateSolutionComponent, data: {title: 'New Solution'}},
   {
     path: ':aid/solutions/:sid',
-    component: SolutionComponent,
-    children: [
-      {path: 'tasks/:task', component: EvaluationModalComponent},
-    ],
+    loadChildren: () => import('./modules/solution/solution.module').then(s => s.SolutionModule),
   },
-  {path: 'courses/create', component: CreateCourseComponent},
-  {path: 'courses/:cid', component: CourseComponent},
-  {path: 'courses/:cid/assignments/:aid', component: CourseComponent},
+  {
+    path: ':aid/edit',
+    loadChildren: () => import('./modules/edit-assignment/edit-assignment.module').then(m => m.EditAssignmentModule),
+  },
+  {path: ':aid', pathMatch: 'full', redirectTo: ':aid/solutions/new'},
+  {
+    path: ':aid',
+    loadChildren: () => import('./modules/assignment/assignment.module').then(m => m.AssignmentModule),
+  },
 ];
 
 @NgModule({
