@@ -1,19 +1,16 @@
-import {DOCUMENT} from '@angular/common';
-import {Component, Inject, OnDestroy, OnInit, ViewChild} from '@angular/core';
-import {ActivatedRoute} from '@angular/router';
-
-import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
+import {Component, OnDestroy, OnInit} from '@angular/core';
+import {ActivatedRoute, Router} from '@angular/router';
 import {forkJoin, of, Subscription} from 'rxjs';
 import {switchMap, tap} from 'rxjs/operators';
 
 import {Marker} from '../../../shared/model/marker';
+import {ToastService} from '../../../toast.service';
 import {UserService} from '../../../user/user.service';
-import {AssignmentService} from '../../services/assignment.service';
-import {CourseService} from '../../services/course.service';
 import Assignment from '../../model/assignment';
 import Course from '../../model/course';
-import {CreateEvaluationDto} from '../../model/evaluation';
 import Solution, {AuthorInfo} from '../../model/solution';
+import {AssignmentService} from '../../services/assignment.service';
+import {CourseService} from '../../services/course.service';
 import {SolutionService} from '../../services/solution.service';
 
 @Component({
@@ -22,25 +19,16 @@ import {SolutionService} from '../../services/solution.service';
   styleUrls: ['./create-solution.component.scss'],
 })
 export class CreateSolutionComponent implements OnInit, OnDestroy {
-  @ViewChild('successModal', {static: true}) successModal;
-
   course?: Course;
   assignment: Assignment;
   solution: string;
   loggedIn = false;
   author: AuthorInfo;
 
-  checking = false;
-  checked = false;
+  status = 'Your solution is checked automatically when you make changes.';
   markers: Marker[] = [];
 
-  id?: string;
-  token?: string;
-  timeStamp?: Date;
-
   submitting: boolean;
-
-  private readonly origin: string;
 
   nextAssignment?: Assignment;
 
@@ -50,12 +38,11 @@ export class CreateSolutionComponent implements OnInit, OnDestroy {
     private courseService: CourseService,
     private assignmentService: AssignmentService,
     private solutionService: SolutionService,
+    private router: Router,
     private route: ActivatedRoute,
-    private modalService: NgbModal,
     private users: UserService,
-    @Inject(DOCUMENT) document: Document,
+    private toastService: ToastService,
   ) {
-    this.origin = document.location.origin;
   }
 
   ngOnInit(): void {
@@ -101,20 +88,9 @@ export class CreateSolutionComponent implements OnInit, OnDestroy {
   getSolution(): Solution {
     return {
       assignment: this.assignment._id!,
-      _id: this.id,
-      token: this.token,
       author: this.author,
       solution: this.solution,
-      timestamp: this.timeStamp,
     };
-  }
-
-  setSolution(solution: Solution): void {
-    this.id = solution._id;
-    this.token = solution.token;
-    this.author = solution.author;
-    this.solution = solution.solution;
-    this.timeStamp = solution.timestamp;
   }
 
   loadDraft(): void {
@@ -128,25 +104,28 @@ export class CreateSolutionComponent implements OnInit, OnDestroy {
 
   check(): void {
     this.saveDraft();
-    this.checking = true;
-
+    this.status = 'Checking...';
     this.solutionService.check({assignment: this.assignment, solution: this.solution}).subscribe(response => {
-      this.checking = false;
-      this.checked = true;
+      this.status = 'Your solution was checked automatically. Don\'t forget to submit when you are done!';
       this.markers = this.assignmentService.lint(response);
+    }, error => {
+      this.status = 'Failed to check your solution automatically: ' + error.error?.message ?? error.message;
     });
   }
 
   submit(): void {
     this.submitting = true;
     this.solutionService.submit(this.getSolution()).subscribe(result => {
-      this.setSolution(result);
-      this.modalService.open(this.successModal, {ariaLabelledBy: 'successModalLabel', size: 'xl'});
+      this.submitting = false;
+      this.toastService.success('Solution', 'Successfully submitted solution');
+      if (this.course && this.nextAssignment) {
+        this.router.navigate(['/assignments', 'courses', this.course._id, 'assignments', this.nextAssignment._id]);
+      } else {
+        this.router.navigate(['/assignments', result.assignment, 'solutions', result._id, 'share']);
+      }
+    }, error => {
+      this.toastService.error('Solution', 'Failed to submit solution', error);
       this.submitting = false;
     });
-  }
-
-  getLink(origin: boolean): string {
-    return `${origin ? this.origin : ''}/assignments/${this.assignment._id}/solutions/${this.id}`;
   }
 }
