@@ -16,7 +16,7 @@ const repoPathMatcher: UrlMatcher = (segments: UrlSegment[]): UrlMatchResult | n
 const routes: Routes = [
   {matcher: repoPathMatcher, component: PageComponent},
   {path: ':repo', redirectTo: ':repo/README.md'},
-  {path: '', component: OverviewComponent},
+  {path: '', component: OverviewComponent, data: {title: 'Docs'}},
 ];
 
 @NgModule({
