@@ -7,8 +7,9 @@ const routes: Routes = [
   {
     path: '',
     component: FourPaneEditorComponent,
+    data: {title: 'Scenario Editor'},
     children: [
-      {path: 'config', component: ConfigComponent},
+      {path: 'config', component: ConfigComponent, data: {title: 'Configure Project'}},
     ],
   },
 ];
