@@ -9,7 +9,6 @@ import 'codemirror/mode/properties/properties';
 import 'codemirror/mode/shell/shell';
 import 'codemirror/mode/yaml/yaml';
 import 'codemirror/addon/mode/simple';
-import 'codemirror/addon/display/autorefresh';
 import 'codemirror/addon/lint/lint';
 import 'codemirror/addon/hint/show-hint';
 import 'codemirror/addon/hint/anyword-hint';
@@ -18,14 +17,17 @@ import * as CodeMirror from 'codemirror';
 import {AppModule} from './app/app.module';
 import {environment} from './environments/environment';
 import {SCENARIO_CODEMIRROR_MODE} from './modes/scenario-codemirror-mode';
+import {TASK_LIST_CODEMIRROR_MODE} from './modes/task-list-codemirror-mode';
 
 CodeMirror.defineSimpleMode('scenario', SCENARIO_CODEMIRROR_MODE);
+CodeMirror.defineSimpleMode('task-list', TASK_LIST_CODEMIRROR_MODE);
 
 import hljs from 'highlight.js/lib/core';
 import java from 'highlight.js/lib/languages/java';
 import groovy from 'highlight.js/lib/languages/groovy';
 import yaml from 'highlight.js/lib/languages/yaml';
 import bnf from 'highlight.js/lib/languages/bnf';
+import json from 'highlight.js/lib/languages/json';
 import {scenario} from './modes/scenario-highlightjs-mode';
 
 hljs.registerLanguage('java', java);
@@ -33,6 +35,7 @@ hljs.registerLanguage('groovy', groovy);
 hljs.registerLanguage('yaml', yaml);
 hljs.registerLanguage('bnf', bnf);
 hljs.registerLanguage('scenario', scenario);
+hljs.registerLanguage(json.name, json);
 
 if (environment.production) {
   enableProdMode();
