@@ -1,14 +1,16 @@
 import {HttpService} from '@nestjs/axios';
 import {Injectable} from '@nestjs/common';
-import {Method} from 'axios';
+import axios, {Method} from 'axios';
+import {createReadStream} from 'fs';
 import {firstValueFrom} from 'rxjs';
-import {AssignmentDocument, Task} from '../assignment/assignment.schema';
+import {Stream} from 'stream';
+import {Entry as ZipEntry, Parse as unzip} from 'unzipper';
+import {AssignmentDocument} from '../assignment/assignment.schema';
 import {AssignmentService} from '../assignment/assignment.service';
 import {environment} from '../environment';
-import {Evaluation, Snippet} from '../evaluation/evaluation.schema';
-import {EvaluationService} from '../evaluation/evaluation.service';
+import {SearchService} from '../search/search.service';
 import {ReadSolutionDto} from '../solution/solution.dto';
-import {Solution, SolutionDocument} from '../solution/solution.schema';
+import {AuthorInfo, Solution, SolutionDocument} from '../solution/solution.schema';
 import {SolutionService} from '../solution/solution.service';
 import {generateToken} from '../utils';
 
@@ -25,20 +27,82 @@ interface SearchResult {
   items: RepositoryInfo[];
 }
 
-interface Issue {
-  number: number;
-  title: string;
-  body: string;
-}
+const MAX_FILE_SIZE = 50 * 1024;
 
 @Injectable()
 export class ClassroomService {
   constructor(
     private assignmentService: AssignmentService,
     private solutionService: SolutionService,
+    private searchService: SearchService,
     private http: HttpService,
-    private evaluationService: EvaluationService,
   ) {
+  }
+
+  async importFiles(id: string, files: Express.Multer.File[]): Promise<ReadSolutionDto[]> {
+    const assignment = await this.assignmentService.findOne(id);
+    if (!assignment) {
+      return [];
+    }
+
+    const result = await this.solutionService.bulkWrite(files.map(file => {
+      const [key, value] = this.parseAuthorInfo(assignment, file.originalname);
+      const author: AuthorInfo = {email: '', name: '', studentId: ''};
+      author[key] = value;
+      const solution: Solution = {
+        assignment: id,
+        solution: '',
+        token: generateToken(),
+        author,
+      };
+      return {
+        updateOne: {
+          filter: {
+            assignment: assignment._id,
+            ['author.' + key]: value,
+          },
+          update: {$setOnInsert: solution},
+          upsert: true,
+        },
+      };
+    }));
+
+    if (assignment.classroom?.codeSearch) {
+      for (let index = 0; index < files.length; index++) {
+        const file = files[index];
+        const stream = createReadStream(file.path);
+        const solution = result.upsertedIds[index];
+        this.importZipStream(stream, id, solution);
+      }
+    }
+
+    return this.solutionService.findAll({_id: {$in: Object.values(result.upsertedIds)}});
+  }
+
+  private parseAuthorInfo(assignment: AssignmentDocument, filename: string): [keyof AuthorInfo, string] {
+    let match: RegExpMatchArray | null;
+    if (/[0-9]/.test(filename) && (match = filename.match(/^([a-zA-Z0-9_.-]+)\.(zip|jar)$/))) {
+      return ['studentId', match[1]];
+    } else if (assignment.classroom && assignment.classroom.prefix && filename.startsWith(assignment.classroom.prefix)) {
+      return ['github', filename.slice(assignment.classroom.prefix.length + 1, -4)];
+    } else {
+      return ['name', filename.slice(0, -4)];
+    }
+  }
+
+  private importZipStream(stream: Stream, assignment: string, solution: string, commit?: string) {
+    stream.pipe(unzip()).on('entry', (entry: ZipEntry) => {
+      if (entry.type !== 'File' || entry.extra.uncompressedSize > MAX_FILE_SIZE) {
+        entry.autodrain();
+        return;
+      }
+      entry.buffer().then(buffer => {
+        const content = buffer.toString('utf-8');
+        let index: number;
+        const path = commit && (index = entry.path.indexOf(commit)) >= 0 ? entry.path.substring(index + commit.length + 1) : entry.path;
+        this.searchService.addFile(assignment, solution, path, content);
+      });
+    });
   }
 
   async importSolutions(id: string, auth: string): Promise<ReadSolutionDto[]> {
@@ -53,7 +117,7 @@ export class ClassroomService {
   }
 
   async importSolutions2(assignment: AssignmentDocument, githubToken: string): Promise<string[]> {
-    const query = `org:${assignment.classroom!.org} ${assignment.classroom!.prefix} in:name`;
+    const query = `org:${assignment.classroom!.org} "${assignment.classroom!.prefix}-" in:name`;
     const repositories: RepositoryInfo[] = [];
     let total = Number.MAX_SAFE_INTEGER;
     for (let page = 1; repositories.length < total; page++) {
@@ -70,14 +134,44 @@ export class ClassroomService {
       const solution = await this.createSolution(assignment, repo, githubToken);
       return {
         updateOne: {
-          filter: {'author.github': solution.author.github},
+          filter: {
+            assignment: assignment._id,
+            'author.github': solution.author.github,
+          },
           update: {$setOnInsert: solution},
           upsert: true,
         },
       };
     }));
     const result = await this.solutionService.bulkWrite(writes);
+
+    if (assignment.classroom?.codeSearch) {
+      const solutions = await this.solutionService.findAll({assignment: assignment.id});
+      for (let solution of solutions) {
+        this.addContentsToIndex(assignment, solution as SolutionDocument, githubToken);
+      }
+    }
+
     return Object.values(result.upsertedIds);
+  }
+
+  private addContentsToIndex(assignment: AssignmentDocument, solution: SolutionDocument, githubToken: string) {
+    const {org, prefix} = assignment.classroom!;
+    const {author: {github}, commit} = solution;
+    if (!github) {
+      return;
+    }
+
+    this.http.get<Stream>(`https://api.github.com/repos/${org}/${prefix}-${github}/zipball/${commit}`, {
+      headers: {
+        Authorization: 'Bearer ' + githubToken,
+      },
+      responseType: 'stream',
+    }).subscribe(response => {
+      this.importZipStream(response.data, assignment.id, solution.id, commit);
+    }, error => {
+      console.error(`Failed to index ${org}/${prefix}-${github}: ${error.message}`);
+    });
   }
 
   private async createSolution(assignment: AssignmentDocument, repo: RepositoryInfo, token: string): Promise<Solution> {
@@ -98,10 +192,17 @@ export class ClassroomService {
     };
   }
 
-  private async getMainCommitSHA(repo: RepositoryInfo, token: string): Promise<string> {
+  private async getMainCommitSHA(repo: RepositoryInfo, token: string): Promise<string | undefined> {
     const url = `https://api.github.com/repos/${repo.full_name}/branches/${repo.default_branch}`;
-    const branch = await this.github<{ commit: { sha: string }; }>('GET', url, token);
-    return branch.commit.sha;
+    try {
+      const branch = await this.github<{ commit: { sha: string }; }>('GET', url, token);
+      return branch.commit.sha;
+    } catch (error) {
+      if (axios.isAxiosError(error) && error.response?.status === 404) {
+        return undefined;
+      }
+      throw error;
+    }
   }
 
   async getGithubToken(auth: string): Promise<string> {
@@ -127,111 +228,5 @@ export class ClassroomService {
       },
     }));
     return data;
-  }
-
-  async exportGithubIssue(assignmentId: string, solutionId: string, auth: string) {
-    const assignment = await this.assignmentService.findOne(assignmentId);
-    if (!assignment || !assignment.classroom || !assignment.classroom.org || !assignment.classroom.prefix) {
-      return;
-    }
-
-    const solution = await this.solutionService.findOne(solutionId);
-    if (!solution || !solution.author.github) {
-      return;
-    }
-
-    const githubToken = await this.getGithubToken(auth);
-    const issue = await this.exportIssue(assignment, solution);
-
-    const baseUrl = `https://api.github.com/repos/${assignment.classroom.org}/${assignment.classroom.prefix}-${solution.author.github}/issues`;
-    const issues = await this.github<Issue[]>('GET', baseUrl, githubToken, {
-      state: 'all',
-    });
-
-    const existing = issues.find(i => i.body.includes(assignmentId));
-    if (existing) {
-      await this.github('PATCH', `${baseUrl}/${existing.number}`, githubToken, {}, issue);
-    } else {
-      await this.github('POST', baseUrl, githubToken, {}, issue);
-    }
-  }
-
-  private async exportIssue(assignment: AssignmentDocument, solution: SolutionDocument): Promise<Omit<Issue, 'number'>> {
-    const {total, sum, tasks} = await this.renderTasks(assignment, solution);
-    const footer = this.renderFooter(assignment, solution);
-
-    return {
-      title: `${assignment.title} (${sum}/${total}P)`,
-      body: `${tasks}\n\n${footer}`,
-    };
-  }
-
-  private async renderTasks(assignment: AssignmentDocument, solution: SolutionDocument) {
-    const evaluations = await this.evaluationService.findAll({
-      assignment: assignment._id,
-      solution: solution._id,
-    });
-    const evaluationRecord: Record<string, Evaluation> = {};
-    for (let evaluation of evaluations) {
-      evaluationRecord[evaluation.task] = evaluation;
-    }
-    const points = this.assignmentService.createPointsCache(assignment.tasks, evaluationRecord);
-    const total = assignment.tasks.reduce((a, c) => c.points > 0 ? a + c.points : 0, 0);
-    const sum = assignment.tasks.reduce((a, c) => a + points[c._id], 0);
-
-    const renderTask = (task: Task, depth: number): string => {
-      const point = points[task._id];
-      const evaluation = evaluationRecord[task._id];
-      const snippets = evaluation ? this.renderSnippets(assignment, solution, evaluation.snippets) : '';
-      if (task.points >= 0) {
-        const headlinePrefix = '#'.repeat(depth + 2);
-        const subTasks = renderSubTasks(task.children, depth + 1);
-        const header = `${headlinePrefix} ${task.description} (${point}/${task.points}P)`;
-        return [header, evaluation?.remark, snippets, subTasks].filter(x => x).join('\n');
-      }
-      if (point === 0) {
-        // do not render deductions that were not given
-        return '';
-      }
-      return `- ${task.description}: ${evaluation.remark} (${point}P)\n${snippets}`;
-    };
-
-    const renderSubTasks = (tasks: Task[], depth: number): string => tasks.map(task => renderTask(task, depth)).join('\n');
-
-    const tasks = renderSubTasks(assignment.tasks, 0);
-    return {total, sum, tasks};
-  }
-
-  private renderSnippets(assignment: AssignmentDocument, solution: SolutionDocument, snippets: Snippet[]) {
-    return snippets.map(snippet => this.renderSnippet(assignment, solution, snippet)).join('\n');
-  }
-
-  private renderSnippet(assignment: AssignmentDocument, solution: SolutionDocument, snippet: Snippet) {
-    const link = `https://github.com/${assignment.classroom!.org}/${assignment.classroom!.prefix}-${solution.author.github}/blob/${solution.commit}/${snippet.file}#L${snippet.from.line + 1}-L${snippet.to.line + 1}`;
-    return `  * ${snippet.comment}: ${link}`;
-  }
-
-  private renderFooter(assignment: AssignmentDocument, solution: SolutionDocument) {
-    const timestamp = new Date();
-    const settings = {
-      'fulibFeedback.apiServer': 'https://fulib.org',
-      'fulibFeedback.assignment.id': assignment._id,
-      'fulibFeedback.solution.id': solution._id,
-      'fulibFeedback.solution.token': solution.token,
-    };
-    return `\
----
-<details>
-<summary>View Evaluations in VSCode</summary>
-
-Copy this to \`.vscode/settings.json\`:
-\`\`\`json
-${JSON.stringify(settings, null, 2)}
-\`\`\`
-
-</details>
-
-<sub>*This issue was created with [fulib.org](https://fulib.org/assignments) on ${timestamp.toLocaleDateString()} at ${timestamp.toLocaleTimeString()} for commit ${solution.commit}.*</sub>
-`;
   }
 }
