@@ -1,7 +1,8 @@
 import {AuthUser, UserToken} from '@app/keycloak-auth';
 import {NotFound, notFound} from '@app/not-found';
-import {Body, Controller, Delete, Get, Headers, Param, Patch, Post} from '@nestjs/common';
+import {Body, Controller, Delete, Get, Headers, MessageEvent, Param, Patch, Post, Sse} from '@nestjs/common';
 import {ApiCreatedResponse, ApiOkResponse, ApiTags} from '@nestjs/swagger';
+import {interval, mapTo, merge, Observable} from 'rxjs';
 import {AssignmentService} from '../assignment/assignment.service';
 import {SolutionAuth} from '../solution/solution-auth.decorator';
 import {CommentAuth} from './comment-auth.decorator';
@@ -34,6 +35,18 @@ export class CommentController {
     const assignment = await this.assignmentService.findOne(assignmentId) ?? notFound(assignmentId);
     const distinguished = this.assignmentService.isAuthorized(assignment, user, assignmentToken);
     return this.commentService.create(assignmentId, solution, dto, distinguished, user?.sub);
+  }
+
+  @Sse('events')
+  @SolutionAuth({forbiddenResponse})
+  events(
+    @Param('assignment') assignment: string,
+    @Param('solution') solution: string,
+  ): Observable<MessageEvent> {
+    return merge(
+      this.commentService.stream(assignment, solution),
+      interval(15000).pipe(mapTo({data: ''})),
+    );
   }
 
   @Get()
