@@ -1,18 +1,20 @@
 import {UserToken} from '@app/keycloak-auth';
-import {Injectable} from '@nestjs/common';
-import {InjectConnection, InjectModel} from '@nestjs/mongoose';
-import {Connection, FilterQuery, Model} from 'mongoose';
+import {Injectable, MessageEvent} from '@nestjs/common';
+import {InjectModel} from '@nestjs/mongoose';
+import {FilterQuery, Model} from 'mongoose';
+import {filter, Observable, Subject} from 'rxjs';
 import {idFilter} from '../utils';
 import {CreateCommentDto, UpdateCommentDto} from './comment.dto';
 import {Comment, CommentDocument} from './comment.schema';
 
 @Injectable()
 export class CommentService {
+  private events$ = new Subject<MessageEvent>();
+
   constructor(
     @InjectModel('comments') private model: Model<Comment>,
-    @InjectConnection() connection: Connection,
   ) {
-    connection.once('connected', () => this.migrate());
+    this.migrate();
   }
 
   async migrate() {
@@ -31,6 +33,14 @@ export class CommentService {
     console.info('Migrated', result.modifiedCount, 'comments');
   }
 
+  stream(assignment: string, solution: string): Observable<MessageEvent> {
+    return this.events$.pipe(filter(({data}) => {
+      console.log(data);
+      const comment = (data as any).comment as Comment;
+      return comment.assignment === assignment && comment.solution === solution;
+    }));
+  }
+
   async create(assignment: string, solution: string, dto: CreateCommentDto, distinguished: boolean, createdBy?: string): Promise<CommentDocument> {
     const comment: Comment = {
       ...dto,
@@ -40,7 +50,9 @@ export class CommentService {
       distinguished,
       timestamp: new Date(),
     };
-    return this.model.create(comment);
+    const created = await this.model.create(comment);
+    this.emit('created', created);
+    return created;
   }
 
   async findAll(where: FilterQuery<Comment> = {}): Promise<CommentDocument[]> {
@@ -52,14 +64,22 @@ export class CommentService {
   }
 
   async update(id: string, dto: UpdateCommentDto): Promise<Comment | null> {
-    return this.model.findOneAndUpdate(idFilter(id), dto, {new: true}).exec();
+    const updated = await this.model.findOneAndUpdate(idFilter(id), dto, {new: true}).exec();
+    updated && this.emit('updated', updated);
+    return updated;
   }
 
   async remove(id: string): Promise<CommentDocument | null> {
-    return this.model.findOneAndDelete(idFilter(id)).exec();
+    const deleted = await this.model.findOneAndDelete(idFilter(id)).exec();
+    deleted && this.emit('deleted', deleted);
+    return deleted;
   }
 
   isAuthorized(comment: Comment, bearerToken: UserToken) {
     return bearerToken.sub === comment.createdBy;
+  }
+
+  private emit(event: string, comment: CommentDocument) {
+    this.events$.next({data: {event, comment}});
   }
 }
