@@ -1,5 +1,5 @@
-import {PartialType} from '@nestjs/swagger';
-import {OmitType} from '@nestjs/swagger';
+import {ApiPropertyOptional, OmitType, PartialType} from '@nestjs/swagger';
+import {IsBoolean, IsOptional} from 'class-validator';
 import {Evaluation} from './evaluation.schema';
 
 export class CreateEvaluationDto extends OmitType(Evaluation, [
@@ -8,7 +8,12 @@ export class CreateEvaluationDto extends OmitType(Evaluation, [
   'createdAt',
   'createdBy',
   'updatedAt',
+  'codeSearch',
 ] as const) {
+  @ApiPropertyOptional()
+  @IsOptional()
+  @IsBoolean()
+  codeSearch?: boolean;
 }
 
 export class UpdateEvaluationDto extends PartialType(CreateEvaluationDto) {
