@@ -10,7 +10,7 @@ async function bootstrap() {
   const prefix = `/api/${environment.version}`;
   app.enableCors();
   app.setGlobalPrefix(prefix);
-  app.useGlobalPipes(new ValidationPipe());
+  app.useGlobalPipes(new ValidationPipe({whitelist: true}));
 
   app.connectMicroservice({
     transport: Transport.NATS,
