@@ -14,7 +14,7 @@ export const environment = {
     bindPrefix: process.env.FULIB_PROJECTS_DATA_DIR || 'data',
     containerImage: process.env.FULIB_PROJECTS_CONTAINER_IMAGE || 'fulib/fulib.org-projects',
     proxyHost: process.env.FULIB_PROJECTS_PROXY_URL || 'http://localhost:8080',
-    network: process.env.FULIB_PROJECTS_NETWORK || 'fulib_projects',
+    network: process.env.FULIB_PROJECTS_NETWORK || 'fuliborg_projects',
   },
   auth: {
     publicKey: `-----BEGIN PUBLIC KEY-----\n${
